Question 2)
1) count the rows using wc -l in the .fam file to determine n
count the rows using wc -l command in the .bim file to determine p

2) -F sets our field operator to "".
$1 designates column 1.
_[$1] looks up $1 in the hash_
++ increments and returns old value
! logical not

Here we are returning values of column 1 that are new to determine chromasomes total in the .bim file, as column 1 represents chromosome number.

To determine how many SNPS in chromosome, which is designated by column 2: pipe results of column 1 (chromosome number) into uniq -c command, to return a count of each type of chromosome.