MAP4 is a gene on chromosome 3, spanning positions 47892180-48130769 bp.
To determine SNPS located within the MAP 4 gene, we use the filter awk to print values where column 1 corresponds to gene 3, and column 4 corresponds to the desired bp range. Instead of prining the results of all the individual SNPS, we pipe the result into the command wc -l to return the number of unique SNPs in the specified range.