hw1.Rmd file header and comments were adjusted to remove redundancy in information. Comments were adjusted for each sub-question to explain the purpose of each line of code. Full sentence responses were added to each sub-question.
Usted knitr::kable to adjust column names of the table to use mathematical font (to allow for subscripts) in column titles.

-For the last part of problem 3: The bash code:

Rscript tablecons.R was removed as it was redundant to the above R code. However, it may still be used to call the tablecons.R script to create a table of means for each distribution at each sample size.

Deleted tablecons.R and replaced with identical tableCons.R for better readability of R script name.