-Updated Header -Added libraries to run, and multiplot function to load
Tidyverse

-Added hw #1 question 2 template.

1) Used count function to count rows of .bim and .fam files.

2) Used group_by to arrange SNPS by chromosome, then used summarise() to count the number in each group.

3) Used filter to set the specifications for a MAP4 gene, then used summarise to count the number meeting the criteria.

4)
-Part 1
We use the select command to create a new subset called snpDef, containing only the coloms we want for a Mendel SNP definition file. We use write_lines() to print the file format version number and number of SNPS as a header into the mendelSnpDef.txt file in the hw2 folder. Then we use write_delim to append our subset snpDef under the header, with comma's as the delimiter.

-Part 2
Similarly we create a subset MendPed with the select command. We use the mutate command to change sex from 1's and 2's into M's and F's. For the IDs we also eliminate all "T2DG"'s with the str_replace_all command. Finally we use write_delim again, to create the appropriately named mendelPedigree.txt file in the hw2 folder.