1) First create a tibble of the highest n count origin airports, then create a tibble of the highest n count destination airports. Determine "busiest" airports by summing the incoming and outgoing n counts to obtain a total n count.
Use innerjoin between the flightCount tibble and the original airports_tbl to retrieve location (lat/long) of those top 10 airports.

We can then write a compressed RDS file and use map_data() and ggmap to plot the busiest airports in the US in 1991( filtered the data by year).

2) Determine the top routes by grouping data by origin and destination. Retrieve the lat/long of the origin and destination airports by using inner_join() once again.

Similarly to the first part we can plot the US map by states, but use geom_curve to connect the geom_point locations of the departure and origin airports.