3 a)
Subset the flights_tbl by origin & destination set as LAX, and filter for years between 1998 and 2008.

Write dataset to a compressed RDS file.

Plot the Flight Dates of LAX, after grouping by date, and counting the totals. We can Use geom_label to label the "prominent" features.

To visualize prominent features, we can filter dates in a smaller range, to get a closer look.