Retrieve a subset from the original tibble, for year, destination, and flight origin, and mutate to season based on month. Count the totals in each season. Plot a stacked bar graph to visualize flights by season
c) Similar Process to part b) except here we select by day of the week instead of season.

d) Filter the flights with origin LAX, and retrieve the longitude and Latitude  of destination location. Next we plot the US map with circles indicating the proportion of all flights from LAX.