4) Create a predictive model for the arrival delay of flights from LAX. Here we are selecting flights from LAX between 2003 and 2007 and using this data to split into a training set and validation set.
With the training set of data we can model a MLR of arrival delay based on the predictive variables distance, departure delay, and weekday. I found these to be the most intuitive results to use for predicting arrival delay of a flight. The summary function returns the model MSE.

Next we can apply the MLR to predict arrival delays of a smaller subset of LAX flights from 2008. Lastly we can visualize the prediction values against the actual arrival delays.